# openapi.yaml
openapi: 3.1.0
info:
  title: Plataforma RH Leoni Cable
  description: >
    Plataforma on-premise de Recursos Humanos para Leoni Cable (Mexico).
    Gestiona solicitudes de empleados (vacaciones, permisos), incidencias,
    actas administrativas, comedor/acceso biometrico, organigrama e
    integracion con nomina TRESS.
  version: 1.0.0
  contact:
    name: Leoni Cable - Equipo de Desarrollo
servers:
  - url: http://localhost:8000
    description: Desarrollo local (Docker)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Clave de API para terminales de comedor
  schemas:
...
    AreaResponse:
      type: object
      required: [area_id, descripcion, estatus_id]
      properties:
        area_id:
          type: integer
        descripcion:
          type: string
        estatus_id:
          type: integer
    SubareaResponse:
      type: object
      required: [subarea_id, area_id, descripcion, estatus_id]
      properties:
        subarea_id:
          type: integer
        area_id:
          type: integer
        descripcion:
          type: string
        estatus_id:
          type: integer
    PuestoResponse:
      type: object
      required: [puesto_id, descripcion, estatus_id]
      properties:
        puesto_id:
          type: integer
        area_id:
          type: [integer, "null"]
        descripcion:
          type: string
        estatus_id:
          type: integer
    CategoriaResponse:
      type: object
      required: [categoria_id, estatus_id]
      properties:
        categoria_id:
          type: integer
        descripcion:
          type: [string, "null"]
        nivel:
          type: [string, "null"]
        estatus_id:
          type: integer
    EstadoEmpleadoResponse:
      type: object
      required: [estado_id, descripcion, estatus_id]
      properties:
        estado_id:
          type: integer
        descripcion:
          type: string
        estatus_id:
          type: integer
    ClasificacionEmpleadoResponse:
      type: object
      required: [clasificacion_id, descripcion, estatus_id]
      properties:
        clasificacion_id:
          type: integer
        descripcion:
          type: string
        significado:
          type: [string, "null"]
        estatus_id:
          type: integer
    RolBrief:
      type: object
      required: [id, nombre]
      properties:
        id:
          type: integer
        nombre:
          type: string
...
    UsuarioResponse:
      type: object
      required: [id, empleado_id, no_empleado, nombre, rol_id, created_at]
      properties:
        id:
          type: integer
        empleado_id:
          type: integer
        no_empleado:
          type: string
        nombre:
          type: string
        email:
          type: [string, "null"]
        rol_id:
          type: integer
        rol:
          $ref: "#/components/schemas/RolBrief"
        estado:
          $ref: "#/components/schemas/EstadoEmpleadoResponse"
        area:
          $ref: "#/components/schemas/AreaResponse"
        subarea:
          $ref: "#/components/schemas/SubareaResponse"
        puesto:
          $ref: "#/components/schemas/PuestoResponse"
        categoria:
          $ref: "#/components/schemas/CategoriaResponse"
        clasificacion:
          $ref: "#/components/schemas/ClasificacionEmpleadoResponse"
        lider_id:
          type: [integer, "null"]
          description: empleado_id del líder inmediato
        centrocosto_id:
          type: [integer, "null"]
        foto:
          type: [string, "null"]
        registro:
          type: [string, "null"]
          format: date
        created_at:
          type: string
          format: date-time
    UsuarioListItem:
      allOf:
        - $ref: "#/components/schemas/UsuarioResponse"
        - type: object
          properties:
            lider_nombre:
              type: [string, "null"]
...
    UsuarioPageResponse:
      type: object
      required: [items, total, page, page_size]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/UsuarioListItem"
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
...
    NivelPuestoCreate:
      type: object
      required: [nombre]
      properties:
        nombre:
          type: string
          minLength: 2
          maxLength: 100
    NivelPuestoUpdate:
      type: object
      required: [nombre]
      properties:
        nombre:
          type: string
          minLength: 2
          maxLength: 100
    NivelPuestoResponse:
      type: object
      required: [id, nombre, activo, created_at, updated_at]
      properties:
        id:
          type: integer
        nombre:
          type: string
        activo:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    NivelPuestoListResponse:
      type: object
      required: [items, total, page, page_size]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/NivelPuestoResponse"
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
...
paths:
...
  /api/v1/empleados:
    get:
      tags: [Empleados]
      summary: Directorio de empleados (paginado)
      operationId: empleados_list
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
        - name: q
          in: query
          description: Busqueda por nombre o numero de empleado
          schema:
            type: string
        - name: area_id
          in: query
          schema:
            type: integer
        - name: puesto_id
          in: query
          description: Permite multiples valores
          schema:
            type: array
            items:
              type: integer
        - name: activo
          in: query
          description: "Solo RH: true=activos, false=inactivos, omitir=todos"
          schema:
            type: boolean
        - name: estatus
          in: query
          description: "Supervisor/Gerente: activo, inactivo, permiso"
          schema:
            type: string
            enum: [activo, inactivo, permiso]
        - name: solo_contratos_por_vencer
          in: query
          description: Contratos que vencen en 30 dias
          schema:
            type: boolean
        - name: solo_sin_lider
          in: query
          description: "Solo RH: empleados activos sin lider asignado (mismo criterio que KPI)"
          schema:
            type: boolean
        - name: solo_sin_email
          in: query
          description: "Solo RH: administrativos activos sin email registrado (mismo criterio que KPI)"
          schema:
            type: boolean
      responses:
        "200":
          description: Pagina de empleados
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsuarioPageResponse"
...
  /api/v1/niveles-puesto:
    get:
      tags: [Talento - Niveles Puesto]
      summary: Listar niveles de puesto (paginado)
      operationId: niveles_puesto_list
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 200
        - name: busqueda
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Pagina de niveles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NivelPuestoListResponse"
    post:
      tags: [Talento - Niveles Puesto]
      summary: Crear nivel de puesto
      operationId: niveles_puesto_create
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NivelPuestoCreate"
      responses:
        "201":
          description: Nivel creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NivelPuestoResponse"
  /api/v1/niveles-puesto/{id}:
    get:
      tags: [Talento - Niveles Puesto]
      summary: Obtener nivel de puesto
      operationId: niveles_puesto_get
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Detalle del nivel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NivelPuestoResponse"
    patch:
      tags: [Talento - Niveles Puesto]
      summary: Actualizar nivel de puesto
      operationId: niveles_puesto_update
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NivelPuestoUpdate"
      responses:
        "200":
          description: Nivel actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NivelPuestoResponse"
    delete:
      tags: [Talento - Niveles Puesto]
      summary: Eliminar nivel de puesto (soft delete)
      operationId: niveles_puesto_delete
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Nivel eliminado

  # ── Talento — Grados Puesto ──────────────────────────
...
  /api/v1/competencias/multihabilidades/puestos:
    get:
      tags: [Talento - Competencias]
      summary: Listar puestos para matriz de multihabilidades
      operationId: competencias_multihabilidades_puestos
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Puestos con competencias configuradas
...